Refactor entities
