+++ src/main/java/com/amt/dflipflop/Entities/authentification/CustomUserDetails.java
@@ -9,7 +9,7 @@
  */
 
 
-package com.amt.dflipflop.Entities;
+package com.amt.dflipflop.Entities.authentification;
 
 /*
 Source : https://www.codejava.net/frameworks/spring-boot/user-registration-and-login-tutorial
