+++ src/main/java/com/amt/dflipflop/WebSecurityConfig.java
@@ -1,21 +1,14 @@
 package com.amt.dflipflop;
 
-import javax.sql.DataSource;
 
-
-import com.amt.dflipflop.Entities.CustomUserDetailsService;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.amt.dflipflop.Services.CustomUserDetailsService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import security.JwtTokenFilter;
